Initialize the TrueTime on method 'onCreate' in Application class.
