+++ app/src/main/java/com/instacart/library/sample/App.java
@@ -1,0 +1,47 @@
+package com.instacart.library.sample;
+
+import android.app.Application;
+import android.util.Log;
+
+import com.instacart.library.truetime.TrueTimeRx;
+
+import java.util.Date;
+
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.observers.DisposableSingleObserver;
+import io.reactivex.schedulers.Schedulers;
+
+public class App extends Application {
+
+    private static final String TAG = App.class.getSimpleName();
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        initRxTrueTime();
+    }
+
+    private void initRxTrueTime() {
+        TrueTimeRx.build()
+                .withConnectionTimeout(31_428)
+                .withRetryCount(100)
+                .withSharedPreferencesCache(getApplicationContext())
+                .withLoggingEnabled(true)
+                .initializeRx("time.google.com")
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribeWith(new DisposableSingleObserver<Date>() {
+                    @Override
+                    public void onSuccess(Date date) {
+                        Log.d(TAG, "Success initialized TrueTime :" + date.toString());
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        Log.e(TAG, "something went wrong when trying to initializeRx TrueTime", e);
+                    }
+                });
+    }
+
+
+}
